Move host/setup files into own dir
